requirements.txt
# JARVIS AI Assistant - Python Dependencies

# AI & ML
google-generativeai>=0.3.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# System & OS Operations
psutil>=5.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Optional but recommended
colorama>=0.4.6  # For colored terminal output
